move stats service to main
